Added a simple Perl module for connecting to search server and parsing its results to Perl data structures, converting nested structures and multiple-valued attributes to arrays.
Also includes the print_tree function which has been originally written
as simple debugging dumper for the parsed query results, but in fact
it's able to dump any complex Perl data structure as long as it's
acyclic.

More to come, including an example (a very simple front-end for the
free version and maybe some more debugging tools).