substantially changed the interface again
I realised that in the last interface I couldn't distinguish between static
and dynamic arrays of length exactly 1.  thinking about how to resolve the
problem led me to a different solution that is as easy to use as before,
but easier to implement, easier to extend by new basic types, and as a
bonus I get a natural interface for static arrays.

I then immediately scratched a duplicate code for parsing dynamic arrays,
because it can be easily unified now.  also, CF_ITEMS does not expect a
parameter now, because they the compiler messes up line numbers when an
error occurs.