Replaced the DARY_ALLOC by a comment explaining when it doesn't work.
I think that we no longer need to worry about this macro, because the only
DARY initializations we needed so far were to an empty array and they are
done automatically now. When the need arises, we can add a simple function
for creating a DARY of a given length.