Added a new mempool primitive mp_spread().
Gradual construction of strings via the growing mempool interface was still
too complicated, so I have introduced a new primitive for this case and also
modified mp_end() to return a pointer to the beginning of the just closed
block (which gets optimized out if unused). This allows to write:

	byte *p = mp_start(pool, 1);
	for (...) {
		p = mp_spread(pool, p, 2);
		*p++ = ...;
	}
	*p++ = 0;
	return mp_end(pool);

Pavel, if you agree, I will merge this into mainline.