A new unit test module.
Explicit input and output files can be defined now, the expected
exit code of the process can be set and, last but not least, test
cases can be named to simplify orientation in long test suites.

The `run' directory is no longer hard-wired, the tester now accepts
--verbose and --rundir switches.

The makefiles now pass --rundir=run and also $TESTERFLAGS, allowing
to turn on the verbose mode at request.