Removed default-install from the generic build system.
The default-install target is very specific for Sherlock and nobody
else seems to use build/install. I have therefore moved this target
back to the top-level Makefile as it was when the world was young.

However, if we compile only the libraries (CONFIG_SHERLOCK is unset),
we use the standard install target identical the one in free/libs/Makefile.