Logging: Log file established log_file() is always dup'ed as stderr.
Previously, stderr was set only at the first open and we forgot to re-dup
it after switching log files.

I have also removed the logic in do_log_reopen() which tried to keep the fd
of the log file stable. Except for having stderr follow the log file, this
was of no use.