Logging: Write a note when some messages fall victim to limiting.
I was considering various shapes of this message. It is tempting to mention
the number of dropped messages, but that has to be done at the time of
an accepted message, which could be too late (the user might already have
wondered about the fate of the missing messages).

Second, I wanted to include a message type, but it would be misleading,
since a single limiter can apply to more types at once with a common bucket.