Heap: Interface cleanup
The previous interfaces was mixing macros, which really perform the
operation, with those where the caller performs the operation and the
macro fixes the heap. This has for example lead to non-systematic side
effects on the `num' variable.

The change is backwards incompatible, all old code should hopefully
fail to compile now.