Opt: Introduced contexts
`struct opt_precomputed' was turning to a option parser context,
so let's admit it. Renamed to `opt_context' and make it passed
as the first argument.

The name of `struct opt_precomputed_option' was a way too unwieldy,
so I renamed it to `struct opt_precomputed'.

Also got rid of several globals variables, but others remain.