Opt: Cleaned up opt-conf
First of all, the names of configuration options are no longer
fixed. Each of these options has its own handler and as long as
OPT_CONF_HOOK is used, the caller is free to define configuration
options with the names and help texts of his choice.

The same applies to --help.

To achieve these goals, the OPT_BEFORE_CONFIG flag has been
introduced.