Table: Clean up column list parsing
First of all, an empty list of columns is a little bit obscure,
but still correct and it should work properly.

table_col_order_by_name() now returns a sensible error message
for non-existent columns.

table_get_col_list() now uses ", " as a separator, which is more
readable. It is documented where the string is allocated from. Also,
it uses the growing string in mempool properly, even in cases when
it is reallocated.

table_start() does not leak memory.