Mainloop: Be benevolent when file_del() is called on a closed fd
I recently wrote a couple of wrappers for connecting other mainloop
interfaces to UCW mainloop. Unfortunately, asking a mainloop to stop
watching a file a moment after close() is a common (mal)practice.
Working around it in the wrapper is hard to do, so I relax the checks
in LibUCW instead.