Logging: Added log_drop_stderr().
When writing the daemonization module, I assumed that it is not
necessary to close stderr, because it will be replaced by a log
file later. Alas, this is not true for general log streams.

I have added a function, which should be called by the daemon
when it is done with initialization of logging.